+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- prevent concurrent edits
+
 ## [0.2.0] - 2025-10-06
 
 ### Changed